Refactor helper functions to fix lint.
